@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.16.2 <3.0.0"
+  sdk: ">=2.17.0 <3.0.0"
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -39,6 +39,7 @@ dependencies:
   shared_preferences: ^2.0.15
   flutter_launcher_icons: ^0.10.0
   provider: ^6.0.4
+  intl: ^0.17.0
 
 dev_dependencies:
   flutter_test:
