@@ -2,7 +2,8 @@ class ApiEndpoint {
   ApiEndpoint._();
 
   static const login = "https://sosty-api.azurewebsites.net/api/User/Login";
-
+  static const signup = "https://sosty-api.azurewebsites.net/api/User/Register";
+  static const getUserById = "https://sosty-api.azurewebsites.net/api/User/GetUserbyID";
   static const getInvestmentsInProgressByInvestor =
       "https://sosty-api.azurewebsites.net/api/Investment/GetInvestmentsInProgressByInvestor";
 }
