@@ -9,7 +9,7 @@ import 'package:sosty/domain/models/investment/gateway/investment_gateway.dart';
 import 'package:sosty/infraestructure/driven_adapter/investment_api/errors/investment_error_api.dart';
 import 'package:sosty/infraestructure/helpers/api_client/api_client.dart';
 import 'package:sosty/infraestructure/helpers/api_client/exception/api_exception.dart';
-import 'package:sosty/ui/common/constants/constants.dart';
+import 'package:sosty/domain/models/common/enums/shared_preferences_enum.dart';
 
 class InvestmentApi extends InvestmentGateway {
   final ApiClient _apiClient = ApiClient();
@@ -19,8 +19,9 @@ class InvestmentApi extends InvestmentGateway {
       String investorID) async {
     final prefs = await SharedPreferences.getInstance();
 
-    final accessToken = prefs.getString(Constants.accessTokenPref) ??
-        Constants.keyPreferenceNotFound;
+    final accessToken =
+        prefs.getString(SharedPreferencesEnum.accessToken.value) ??
+            SharedPreferencesEnum.keyNotFound.value;
 
     final response = await _apiClient.get(
       ApiEndpoint.getInvestmentsInProgressByInvestor,
@@ -36,7 +37,7 @@ class InvestmentApi extends InvestmentGateway {
     } else {
       throw ApiException(
         ErrorItem(
-          domain: ApiEndpoint.login,
+          domain: ApiEndpoint.getInvestmentsInProgressByInvestor,
           reason: response.body,
           message: InvestmentErrorApi.getInvestmentsInProgressByInvestorFailed,
         ),
