@@ -1,21 +1,21 @@
 import 'package:flutter/material.dart';
-import 'package:sosty/ui/common/constants/constants.dart';
 import 'package:sosty/ui/common/styles/styles.dart';
+import 'package:sosty/ui/common/enums/assets_enum.dart';
 
 class NavbarDetail extends StatelessWidget {
   const NavbarDetail({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
-    return const SliverAppBar(
+    return SliverAppBar(
       pinned: false,
       snap: true,
       floating: true,
       centerTitle: true,
       backgroundColor: Styles.primaryColor,
-      leading: BackButton(color: Colors.white),
+      leading: const BackButton(color: Colors.white),
       title: Image(
-        image: AssetImage(Constants.assetLogo),
+        image: AssetImage(AssetsEnum.logo.value),
         height: 50,
       ),
     );
